refactor: split TransformerFlags, TransformerCfg and compile-time, macro dependend TransformerConfiguration to separate compilation units
