+++ chimney/src/main/scala/io/scalaland/chimney/internal/macros/Model.scala
@@ -1,7 +1,5 @@
 package io.scalaland.chimney.internal.macros
 
-import io.scalaland.chimney.internal.TransformerConfiguration
-
 import scala.reflect.macros.blackbox
 
 trait Model extends TransformerConfiguration {
